% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{compare_2d_map_correlations}
\alias{compare_2d_map_correlations}
\title{Compare correlations of two sets of spatial maps}
\usage{
compare_2d_map_correlations(map1, map2, Q = NULL, cormethod = "pearson")
}
\arguments{
\item{map1}{A Q x V matrix, with each row corresponding to a component}

\item{map2}{A Q x V matrix, with each row corresponding to a component
to the number of independent components that you plan to include in the model.}

\item{Q}{Optional parameter giving the number of components This will be used
to fixed the orientation of map1 and map2 if needed}

\item{cormethod}{technique used to calculate the correlation. Default is "pearson"}
}
\value{
\itemize{
  \item correlation_unsorted - The raw correlation matrix between the two sets
of maps
  \item sorted_order - A vector that can be used to sort map1 to be most aligned with map2
  \item correlation_sorted - The correlation between the two sets of maps after map1 has been sorted.
For this result, the largest correlation should be along the diagonal.
}
}
\description{
Compare correlations of two sets of spatial maps
}
\details{
`compare_2d_map_correlations` takes two sets of spatial maps as arguments and
returns the correlation between them as well as information about sorting
them to be as aligned as possible.
}
