% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{create_sparsebayes_timing_log}
\alias{create_sparsebayes_timing_log}
\title{Internal function that generates a dataframe with all the timing information
to be filled out}
\usage{
create_sparsebayes_timing_log()
}
\description{
Internal function that generates a dataframe with all the timing information
to be filled out
}
\keyword{internal}
