% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{draw_shape}
\alias{draw_shape}
\title{Function to draw a spatial map with digits as the activation region}
\usage{
draw_shape(multvec, xoff, yoff, block_size, block_thick, imgsize)
}
\description{
Function to draw a spatial map with digits as the activation region
}
\keyword{internal}
