% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{evaluate_effect_discovery}
\alias{evaluate_effect_discovery}
\title{Get TPR, FDR, etc}
\usage{
evaluate_effect_discovery(pvalues, true_effect_indicators)
}
\arguments{
\item{pvalues}{p-values for the method of interest}

\item{true_effect_indicators}{Matrix with 1 where truth is non-zero and 0 otherwise}
}
\value{
A list containing standard evaluation quantities
}
\description{
Get TPR, FDR, etc
}
