% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{export_to_cluster}
\alias{export_to_cluster}
\title{Sends data to the specific parallel worker}
\usage{
export_to_cluster(cluster, var_name, variable)
}
\description{
Sends data to the specific parallel worker
}
\keyword{internal}
