% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{export_worker_state}
\alias{export_worker_state}
\title{saves a set of objects on a parallel worker}
\usage{
export_worker_state(fname)
}
\description{
saves a set of objects on a parallel worker
}
\keyword{internal}
