% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{generate_covariate_values}
\alias{generate_covariate_values}
\title{Internal function to generate a set of covariate values}
\usage{
generate_covariate_values(
  N,
  n_discrete,
  n_continuous,
  interactions,
  coding = "reference"
)
}
\description{
Internal function to generate a set of covariate values
}
\keyword{internal}
