% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{generate_digits_example_data}
\alias{generate_digits_example_data}
\title{Generate Digits Example Data}
\usage{
generate_digits_example_data(
  N,
  Q,
  sigma_sq_y,
  sigma_sq_q,
  population_map_mean,
  population_map_var,
  beta_mean,
  beta_var,
  n_continuous_cov = 1,
  n_discrete_cov = 0,
  interactions = NA,
  slice_width = 75,
  time_points = 100
)
}
\arguments{
\item{N}{the number of subjects}

\item{Q}{the number of independent components}

\item{sigma_sq_y}{scalar variance of the time series noise}

\item{sigma_sq_q}{vector of length Q containing the between subject variances in the components}

\item{population_map_mean}{scalar average for the population level map activation regions}

\item{population_map_var}{scalar variance for the population level map activation regions}

\item{beta_mean}{scalar average for the population level map activation regions}

\item{beta_var}{scalar variance for the population level map activation regions}

\item{n_continuous_cov}{number of continuous covariates in the data. Default is 1}

\item{n_discrete_cov}{number of categorical covariates in the data (e.g. treatment group). Default is 0.}

\item{interactions}{matrix of interactions. Each row corresponds to a single
interaction. The columns are 1 if the variable is included in that interaction
and 0 otherwise. Default is NA, which means no interactions in the model.}

\item{slice_width}{The number of voxels along one side of the slice. The total
number of voxels in the brain will be slice_width squared.}

\item{time_points}{the number of time points to generate for each subject}
}
\value{
\itemize{
  \item data_directory - where the data were stored
  \item A - the mixing matrix for each subject
  \item S0 - the population level maps
  \item Si - the subject level maps
  \item beta - the covariate effect spatial maps
  \item Y - Q x V time series data for each subject
  \item X - matrix with the covariates for each subject
  \item sigma_sq_q - same as argument
  \item sigma_sq_y - same as argument
}
}
\description{
Generate Digits Example Data
}
\details{
This function will generate a dataset for demonstrating the SparseBayes ICA
approach. The data consist of a single slice, with activation regions in the
shape of digits from 0-9. Covariate effects are placed within these activation
regions.
}
