% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{generate_mixing_matrix}
\alias{generate_mixing_matrix}
\title{Internal function to generate a random Q x Q orthonormal matrix for a set of N subjects}
\usage{
generate_mixing_matrix(N, Q)
}
\description{
Internal function to generate a random Q x Q orthonormal matrix for a set of N subjects
}
\keyword{internal}
