% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{generate_Si}
\alias{generate_Si}
\title{Function to draw a subject specific maps for the digits example}
\usage{
generate_Si(S0, beta, X, sigma_sq_q)
}
\description{
Function to draw a subject specific maps for the digits example
}
\keyword{internal}
