% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{generate_time_courses}
\alias{generate_time_courses}
\title{Function to create fMRI time courses for the digits example}
\usage{
generate_time_courses(A, Si, sigma_sq_y)
}
\description{
Function to create fMRI time courses for the digits example
}
\keyword{internal}
