% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{get_shape}
\alias{get_shape}
\title{Function to draw a spatial map with numbers as the activation region}
\usage{
get_shape(number, xoff, yoff, block_size = 10, block_thick = 5, imgsize = 100)
}
\description{
Function to draw a spatial map with numbers as the activation region
}
\keyword{internal}
