% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{initialize_sparsebayes_nodes}
\alias{initialize_sparsebayes_nodes}
\title{Does the work of splitting all needed variables across the different
parallel workers}
\usage{
initialize_sparsebayes_nodes(
  cl,
  voxel_combinations,
  subject_combinations,
  Yin,
  Xin,
  guess,
  angle_range,
  sin_theta,
  cos_theta,
  S0_DPM
)
}
\description{
Does the work of splitting all needed variables across the different
parallel workers
}
\keyword{internal}
