% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{load_mask}
\alias{load_mask}
\title{Load a brain mask from a Nifti file}
\usage{
load_mask(mask_path)
}
\arguments{
\item{mask_path}{The file path to the mask file, which should be in nifti format.}
}
\value{
A list containing the following:
\itemize{
  \item mask - the nifti object for the mask
  \item valid_voxels - a list containing the linear indices of the voxels that
  are in the masked region
  \item V - the total number of voxels in the brain mask
}
}
\description{
Load a brain mask from a Nifti file
}
