% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_data_formatting.R
\name{matrix_to_nifti}
\alias{matrix_to_nifti}
\title{Take a matrix and use it to create a brain volume}
\usage{
matrix_to_nifti(data, mask_info, reference_image = NULL)
}
\arguments{
\item{data}{A data matrix with V elements}

\item{mask_info}{The output from using \code{\link{load_mask}} on a nifti file
containing a brain mask for a brain with V voxels.}

\item{reference_image}{Imaging containing other header information. Default is NULL, which
results in the mask file header being used.}
}
\value{
A nifti volume
}
\description{
Take a matrix and use it to create a brain volume
}
\details{
`matrix_to_nifti` takes a matrix with V (number of voxels) elements and places
the values into a brain. The mappings from the matrix to the brain are provided by
the mask_info object, which can be obtained using \code{\link{load_mask}}.
}
\seealso{
[nifti_to_matrix()]
}
