% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_data_formatting.R
\name{nifti_to_matrix}
\alias{nifti_to_matrix}
\title{Flatten a Nifti image to a matrix}
\usage{
nifti_to_matrix(data, valid_voxels)
}
\arguments{
\item{data}{4 dimensional Nifti data for a subject}

\item{valid_voxels}{A list of the linear indices of the voxels in the brain mask.
This can be obtained using \code{\link{load_mask}}}
}
\value{
A T x V data matrix
}
\description{
Flatten a Nifti image to a matrix
}
\details{
`nifti_to_matrix` takes a subject's fMRI volumes from a NIFTI file and converts them
to a T x V data matrix, where T is the number of volumes and V is the number of
voxels in the brain.
}
\examples{
library(SparseBayesICA)

# Get the filepath to the example data mask (provided in this package)
mask_path  <- system.file("extdata", "example_mask.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the mask
mask_info  <- load_mask(mask_path)

# Filepath for one of the example subjects provided in this package
nifti_path <- system.file("extdata", "example_subject_1.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the data
nifti_data <- readNifti(nifti_path)

# Apply the mask to the data to get a T x V data matrix instead of 4D data
masked_data <- nifti_to_matrix(nifti_data, mask_info$valid_voxels)

}
\seealso{
[load_mask()]
}
