% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_and_initialization.R
\name{obtain_initial_guess_sparsebayes}
\alias{obtain_initial_guess_sparsebayes}
\title{Use the population level components to obtain a set of initial values for all
SparseBayes ICA model parameters}
\usage{
obtain_initial_guess_sparsebayes(data, S0, X, n_initial_cluster = 10)
}
\arguments{
\item{data}{A list containing the prewhitened data for each subject.}

\item{S0}{The initial guess for the population level parameters}

\item{X}{A matrix of covariates with N rows and P columns}

\item{n_initial_cluster}{The number of initial clusters for the DPM. Default is 10.}
}
\value{
A list object containing the initial parameters for SparseBayes ICA
}
\description{
`obtain_initial_guess_sparsebayes()` Uses the population level initial values
generated from `obtain_s0_guess_fastica()` to obtain initial parameter settings
for all SparseBayes ICA model parameters.
}
