% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{partition_into_chunks}
\alias{partition_into_chunks}
\title{Partitions the input data into different chunks to send to the different worker
nodes in parallel sparsebayes. This returns the corresponding index sets}
\usage{
partition_into_chunks(n_node, combinations)
}
\description{
Partitions the input data into different chunks to send to the different worker
nodes in parallel sparsebayes. This returns the corresponding index sets
}
\keyword{internal}
