% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_and_initialization.R
\name{PCA_dimension_reduction}
\alias{PCA_dimension_reduction}
\title{Dimensionality reduction of a set of fMRI time courses to a set of principal components}
\usage{
PCA_dimension_reduction(data, nPC)
}
\arguments{
\item{data}{A T x V dataset where T is the number of fMRI volumes and V is the
number of voxels.}

\item{nPC}{The requested number of principal components}
}
\value{
A nPC x V data matrix with rows containing the principal components
}
\description{
`PCA_dimension_reduction()` takes a set of fMRI time courses and reduces it
to a set of `nPC` orthogonal time courses. This is useful for creating
dimension reduced subject and group level data for temporal concatenation
group ICA, which can be used to obtain starting values for SparseBayes ICA
if desired.
}
\details{
Let \eqn{Y} be the fMRI time courses with T rows and V columns. After centering,
we can apply a spectral decomposition to the \eqn{T \times T} data covariance
to obtain:
\deqn{Cov(Y) = EDE'}
This function then returns \eqn{E^* D^*}, where \eqn{E^*} consists of the
first `nPC` columns of \eqn{E} and \eqn{D^*} is a diagonal matrix with the
`nPC` largest eigenvalues of \deqn{Cov(Y)}.
}
\examples{
library(SparseBayesICA)

# Get the filepath to the example data mask (provided in this package)
mask_path  <- system.file("extdata", "example_mask.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the mask
mask_info  <- load_mask(mask_path)

# Filepath for one of the example subjects provided in this package
nifti_path <- system.file("extdata", "example_subject_1.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the data
nifti_data <- readNifti(nifti_path)

# Apply the mask to the data to get a T x V data matrix instead of 4D data
masked_data <- nifti_to_matrix(nifti_data, mask_info$valid_voxels)

# Perform dimension reduction to 10 principal components
subject_PCA <- PCA_dimension_reduction(masked_data, nPC = 10)

}
