% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_and_initialization.R
\name{preprocess_subject}
\alias{preprocess_subject}
\title{Preprocess a subject's fMRI time courses}
\usage{
preprocess_subject(data, Q)
}
\arguments{
\item{data}{A T x V dataset where T is the number of fMRI volumes and V is the
number of voxels.}

\item{Q}{The requested dimension of the prewhitened data. This should be equal
to the number of independent components that you plan to include in the model.}
}
\value{
A list containing the prewhitnened data and the eigenvalues of the
original scale data covariance
}
\description{
Preprocess a subject's fMRI time courses
}
\examples{
library(SparseBayesICA)

# Get the filepath to the example data mask (provided in this package)
mask_path  <- system.file("extdata", "example_mask.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the mask
mask_info  <- load_mask(mask_path)

# Filepath for one of the example subjects provided in this package
nifti_path <- system.file("extdata", "example_subject_1.nii",
                           package = "SparseBayesICA", mustWork = TRUE)

# Load the data
nifti_data <- readNifti(nifti_path)

# Apply the mask to the data to get a T x V data matrix instead of 4D data
masked_data <- nifti_to_matrix(nifti_data, mask_info$valid_voxels)

# Prewhiten this subject's data to Q = 5 whitened time points
prewhitened_results <- preprocess_subject(masked_data, Q = 5)

}
\seealso{
[load_mask()], [PCA_dimension_reduction()]
}
