% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_and_initialization.R
\name{random_initialization}
\alias{random_initialization}
\title{Get an initial guess using a random set of starting values}
\usage{
random_initialization(data, X, n_initial_cluster = 10)
}
\arguments{
\item{data}{A list containing the prewhitened data for each subject.}

\item{X}{A matrix of covariates with N rows and P columns}

\item{n_initial_cluster}{The number of initial clusters for the DPM. Default is 10.}
}
\value{
A list object containing the initial parameters for SparseBayes ICA
}
\description{
`random_initialization()` Generates a random set of starting values for
SparseBayes ICA. In general, this should be avoided in favor of using TCGICA
and dual regression.
}
