% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_matrix_wrapper.R
\name{sample_mixing_matrix_inplace}
\alias{sample_mixing_matrix_inplace}
\title{wrapper for mixing matrix function}
\usage{
sample_mixing_matrix_inplace(
  A,
  YSi,
  YSiOmega,
  Omega,
  YYt,
  angle_range,
  sin_theta,
  cos_theta
)
}
\description{
wrapper for mixing matrix function
}
\keyword{internal}
