% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SparseBayesICA}
\alias{SparseBayesICA}
\title{SparseBayesICA}
\description{
Package implementing codes for independent component analysis of multi-subject
fMRI data with sparse estimation of covariate effects.
}
\author{
Joshua Lukemire <joshua.lukemire@emory.edu>
}
