% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseBayesICASerial.R
\name{SparseBayesICASerial}
\alias{SparseBayesICASerial}
\title{Internal function running the MCMC when a parallel environment is NOT used
this is called by fit_SparseBayesICA, and is not a public function}
\usage{
SparseBayesICASerial(
  Y,
  X,
  initial_values,
  nburnin = 5000,
  nmcmc = 5000,
  concentration = 1,
  print_freq = 100
)
}
\description{
Internal function running the MCMC when a parallel environment is NOT used
this is called by fit_SparseBayesICA, and is not a public function
}
\keyword{internal}
