% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{update_cluster_matrix_variable_rowspec}
\alias{update_cluster_matrix_variable_rowspec}
\title{updates a matrix with a specific index set on the corresponding parallel worker}
\usage{
update_cluster_matrix_variable_rowspec(
  cluster,
  var_name,
  variable,
  row_index_list_name
)
}
\description{
updates a matrix with a specific index set on the corresponding parallel worker
}
\keyword{internal}
