% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helper_functions.R
\name{update_cluster_vector_variable}
\alias{update_cluster_vector_variable}
\title{updates a vector on the corresponding parallel worker}
\usage{
update_cluster_vector_variable(cluster, var_name, variable)
}
\description{
updates a vector on the corresponding parallel worker
}
\keyword{internal}
