% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_digits_map_generation.R
\name{write_data_to_TxV}
\alias{write_data_to_TxV}
\title{Function to create fMRI time courses for the digits example}
\usage{
write_data_to_TxV(Y, nT, output_directory, mask_info, use_Di = FALSE)
}
\description{
Function to create fMRI time courses for the digits example
}
\keyword{internal}
