% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{load_mask}
\alias{load_mask}
\title{Load a specified brain mask}
\usage{
load_mask(mask_path)
}
\arguments{
\item{mask_path}{The file path to the mask file, which should be in nifti format.}
}
\value{
An R list containing the indices of the voxels in the brain, the total number of voxels,
and the mask nifti object.
}
\description{
Load a specified brain mask
}
